init salesorder

+++ src/main/java/com/smartshop/eshop/repository/CountryRepository.java
@@ -13,6 +13,6 @@ public interface CountryRepository extends JpaRepository<Country, Long> {
 	Country findByIsoCode(String code);
 
 	@Query("select c from Country c left join fetch c.descriptions cd where cd.language.id=?1")
-	List<Country> listByLanguage(Integer id);
+	List<Country> listByLanguage(Long id);
 
 }
