+++ src/main/java/com/smartshop/eshop/repository/ZoneRepository.java
@@ -12,9 +12,9 @@ public interface ZoneRepository extends JpaRepository<Zone, Long> {
 	Zone findByCode(String code);
 
 	@Query("select z from Zone z left join fetch z.descriptions zd where zd.language.id=?1")
-	List<Zone> listByLanguage(Integer id);
+	List<Zone> listByLanguage(Long id);
 
 	@Query("select z from Zone z left join fetch z.descriptions zd join fetch z.country zc where zc.isoCode=?1 and zd.language.id=?2")
-	List<Zone> listByLanguageAndCountry(String isoCode, Integer languageId);
+	List<Zone> listByLanguageAndCountry(String isoCode, Long languageId);
 
 }
